Fixed incomplete file download.
Fixed download problem where files that were sent from the Teensy4.x to the PC were incomplete. File size was smaller than the original file.  The problem was in the doRetrieve() function. The data.write() buffer was being overwritten before the last transfer was finished. Limited the read size to space left in data.write() buffer if it was smaller than FTP_BUF_SIZE.